Enforce setting privateNetworking for fully-private clusters
privateNetworking for both self-managed and EKS-managed nodegroups must be set explicitly. It's an error to leave privateNetworking unset, as it's not possible to create public nodegroups in a fully-private cluster as there's no public subnet. Rather than enabling this field by default for a fully-private cluster, we require users to explicitly set it to make the behaviour explicit and avoid confusion.